// frontend/src/api/models/update-meeting-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateMeetingDTO
 */
export interface UpdateMeetingDTO {
    /**
     * 
     * @type {string}
     * @memberof UpdateMeetingDTO
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateMeetingDTO
     */
    'meetingStart'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateMeetingDTO
     */
    'meetingEnd'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateMeetingDTO
     */
    'location'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateMeetingDTO
     */
    'meetingStatus'?: UpdateMeetingDTOMeetingStatusEnum;
}

export const UpdateMeetingDTOMeetingStatusEnum = {
    Pending: 'PENDING',
    Confirmed: 'CONFIRMED',
    Cancelled: 'CANCELLED'
} as const;

export type UpdateMeetingDTOMeetingStatusEnum = typeof UpdateMeetingDTOMeetingStatusEnum[keyof typeof UpdateMeetingDTOMeetingStatusEnum];